fix: Use late static binding in AbstractProvider (#123)
## This PR
Switches from self::$NAME to static::$NAME in the getMetadata method of
the AbstractProvider class. This change is to ensure the name of the
extending provider is used in the metadata. Previously, when extending
from AbstractProvider, the metadata name was always AbstractProvider.

### How to test
Create a class that extends AbstractProvider, give it a different name,
then call "getMetadata" and see that the name is now the name of the
extended class and not "AbstractProvider"

Signed-off-by: Andrew Menino-Barlow <andrew.barlow@gmail.com>